Add support for ignoring non filepath based imports.
This replaces import statements of the form:
```javascript
import foo from 'bar';
```
where bar is not a relative or absolute file path, with nothing.

If bar starts with . or / it is importing from a file on disk and the
existing behaviour of replacing it with a global is retained.

Fixes #6